Rework simulator to account for latency
The old interpreter didn't account for latency and was built on the
assumption that potential race conditions do not occur in real circuits.
However, it turns out many circuits _do_ rely on the switching latency
of gates for correct behaviour.

The new simulator uses a more conventional approach where changes in the
circuit are propagated by updating all affected nodes in the same step.
Separate read & write buffers are used to make latency work as expected.